[PC-1284] Make openagenda localprovider test non required
